refactor: Switch to hooks
