+++ template/src/routes/home/index.js
@@ -1,5 +1,5 @@
 import { h } from 'preact';
-import style from './style';
+import style from './style.css';
 
 const Home = () => (
 	<div class={style.home}>
