+++ template/src/routes/profile/index.js
@@ -1,47 +1,30 @@
-import { h, Component } from 'preact';
-import style from './style';
+import { h } from 'preact';
+import {useEffect, useState} from "preact/hooks";
+import style from './style.css';
 
-export default class Profile extends Component {
-	state = {
-		time: Date.now(),
-		count: 10
-	};
+// Note: `user` comes from the URL, courtesy of our router
+const Profile = ({ user }) => {
+	const [time, setTime] = useState(Date.now());
+	const [count, setCount] = useState(10);
 
-	// update the current time
-	updateTime = () => {
-		this.setState({ time: Date.now() });
-	};
+	useEffect(() => {
+		setInterval(() => setTime(Date.now()), 1000);
+	}, []);
 
-	increment = () => {
-		this.setState({ count: this.state.count+1 });
-	};
+	return (
+		<div class={style.profile}>
+			<h1>Profile: {user}</h1>
+			<p>This is the user profile for a user named { user }.</p>
 
-	// gets called when this route is navigated to
-	componentDidMount() {
-		// start a timer for the clock:
-		this.timer = setInterval(this.updateTime, 1000);
-	}
+			<div>Current time: {new Date(time).toLocaleString()}</div>
 
-	// gets called just before navigating away from the route
-	componentWillUnmount() {
-		clearInterval(this.timer);
-	}
-
-	// Note: `user` comes from the URL, courtesy of our router
-	render({ user }, { time, count }) {
-		return (
-			<div class={style.profile}>
-				<h1>Profile: {user}</h1>
-				<p>This is the user profile for a user named { user }.</p>
-
-				<div>Current time: {new Date(time).toLocaleString()}</div>
-
-				<p>
-					<button onClick={this.increment}>Click Me</button>
-					{' '}
-					Clicked {count} times.
-				</p>
-			</div>
-		);
-	}
+			<p>
+				<button onClick={() => setCount((count) => ++count)}>Click Me</button>
+				{' '}
+				Clicked {count} times.
+			</p>
+		</div>
+	);
 }
+
+export default Profile;
